ESTHER's TOKEN
RESEARCH ON ERC20 TOKENS
ERC-20 stands for "Ethereum Request for Comment 20," and it represents a standard interface for fungible tokens on the Ethereum blockchain. These tokens are digital assets with a set value, just like a dollar or a euro, and can be used as a unit of value within the Ethereum ecosystem.

The ERC-20 standard outlines a set of rules and functions that Ethereum-based tokens must follow to be considered ERC-20 compliant. These rules include a specific set of methods for transferring tokens, querying balances, and other basic functionalities, allowing different tokens to be easily exchanged and interacted with on the Ethereum network.